Split data analysis execution API from chat API
