+++ lib/galaxy/webapps/galaxy/api/data_analysis_chat_execution.py
@@ -1,0 +1,184 @@
+"""Data analysis chat execution endpoints for chat-backed Pyodide flows."""
+
+from typing import Optional
+
+from fastapi import (
+    File,
+    Form,
+    HTTPException,
+    Query,
+    status,
+    UploadFile,
+    WebSocket,
+)
+from starlette.responses import StreamingResponse
+
+from galaxy.exceptions import (
+    AuthenticationFailed,
+    InternalServerError,
+    ItemAccessibilityException,
+    MalformedId,
+    ObjectNotFound,
+    RequestParameterInvalidException,
+    RequestParameterMissingException,
+)
+from galaxy.managers.context import ProvidesHistoryContext
+from galaxy.managers.data_analysis_chat_artifacts import DataAnalysisChatArtifactsManager
+from galaxy.managers.data_analysis_chat_datasets import DataAnalysisChatDatasetsManager
+from galaxy.managers.data_analysis_chat_execution import DataAnalysisChatExecutionManager
+from galaxy.managers.session import GalaxySessionManager
+from galaxy.managers.users import UserManager
+from galaxy.model import User
+from galaxy.schema.agents import UploadedArtifact
+from galaxy.schema.fields import DecodedDatabaseIdField
+from galaxy.schema.schema import PyodideResultPayload
+from galaxy.security.idencoding import IdEncodingHelper
+from galaxy.webapps.galaxy.api import (
+    depends,
+    DependsOnTrans,
+    DependsOnUser,
+    get_app,
+    Router,
+)
+from galaxy.work.context import WorkRequestContext
+
+router = Router(tags=["chat"])
+
+
+def _get_websocket_user(websocket: WebSocket, app) -> Optional[User]:
+    session_cookie = websocket.cookies.get("galaxysession")
+    if session_cookie:
+        security = app[IdEncodingHelper]
+        session_manager = app[GalaxySessionManager]
+        try:
+            session_key = security.decode_guid(session_cookie)
+        except MalformedId:
+            session_key = None
+        if session_key:
+            galaxy_session = session_manager.get_session_from_session_key(session_key)
+            if galaxy_session and galaxy_session.user:
+                return galaxy_session.user
+
+    user_manager = app[UserManager]
+    api_key = websocket.query_params.get("key") or websocket.headers.get("x-api-key")
+    if api_key:
+        try:
+            return user_manager.by_api_key(api_key)
+        except AuthenticationFailed:
+            return None
+
+    authorization = websocket.headers.get("authorization")
+    if authorization and authorization.startswith("Bearer "):
+        access_token = authorization.removeprefix("Bearer ").strip()
+        if access_token:
+            return user_manager.by_oidc_access_token(access_token)
+
+    return None
+
+
+@router.cbv
+class DataAnalysisChatExecutionAPI:
+    data_analysis_chat_artifacts_manager: DataAnalysisChatArtifactsManager = depends(DataAnalysisChatArtifactsManager)
+    data_analysis_chat_datasets_manager: DataAnalysisChatDatasetsManager = depends(DataAnalysisChatDatasetsManager)
+    data_analysis_chat_execution_manager: DataAnalysisChatExecutionManager = depends(DataAnalysisChatExecutionManager)
+
+    @router.post("/api/chat/exchange/{exchange_id}/artifacts")
+    async def upload_pyodide_artifact(
+        self,
+        exchange_id: DecodedDatabaseIdField,
+        file: UploadFile = File(...),
+        name: Optional[str] = Form(default=None),
+        mime_type: Optional[str] = Form(default=None),
+        size: Optional[int] = Form(default=None),
+        trans: ProvidesHistoryContext = DependsOnTrans,
+        user: User = DependsOnUser,
+    ) -> UploadedArtifact:
+        """Persist an artifact generated by the Pyodide worker as a history dataset."""
+
+        raw_bytes = await file.read()
+        artifact_name = name or file.filename or "artifact"
+        artifact_mime = mime_type or file.content_type or "application/octet-stream"
+        try:
+            return self.data_analysis_chat_artifacts_manager.create_artifact(
+                exchange_id=exchange_id,
+                trans=trans,
+                raw_bytes=raw_bytes,
+                name=artifact_name,
+                mime_type=artifact_mime,
+                size=size,
+            )
+        except ObjectNotFound as exc:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(exc)) from exc
+        except RequestParameterInvalidException as exc:
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(exc)) from exc
+        except InternalServerError as exc:
+            raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=str(exc)) from exc
+
+    @router.post("/api/chat/exchange/{exchange_id}/pyodide_result")
+    async def submit_pyodide_result(
+        self,
+        exchange_id: DecodedDatabaseIdField,
+        payload: PyodideResultPayload,
+        trans: ProvidesHistoryContext = DependsOnTrans,
+        user: User = DependsOnUser,
+    ) -> dict[str, object]:
+        """Persist results from client-side Pyodide execution and trigger follow-up reasoning."""
+
+        try:
+            return await self.data_analysis_chat_execution_manager.submit_pyodide_result(exchange_id, payload, trans, user)
+        except ObjectNotFound:
+            return {"message": "Chat exchange not found"}
+
+    @router.get("/api/chat/datasets/{dataset_id}/download")
+    async def download_dataset_for_execution(
+        self,
+        dataset_id: str,
+        token: str = Query(..., description="Signed dataset download token"),
+        trans: ProvidesHistoryContext = DependsOnTrans,
+        user: User = DependsOnUser,
+    ):
+        """Stream a history dataset referenced by a signed execution token."""
+
+        try:
+            prepared_download = self.data_analysis_chat_datasets_manager.prepare_download(trans, user, dataset_id, token)
+        except RequestParameterMissingException as exc:
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(exc)) from exc
+        except ItemAccessibilityException as exc:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=str(exc)) from exc
+        except ObjectNotFound as exc:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(exc)) from exc
+
+        safe_name = prepared_download.display_name.replace("\\", "").replace('"', "")
+        headers = {
+            "Content-Disposition": f'attachment; filename="{safe_name}"',
+            "Content-Length": str(prepared_download.content_length),
+        }
+
+        def iter_file():
+            with open(prepared_download.file_path, "rb") as handle:
+                while True:
+                    chunk = handle.read(65536)
+                    if not chunk:
+                        break
+                    yield chunk
+
+        return StreamingResponse(iter_file(), media_type=prepared_download.mime_type, headers=headers)
+
+
+@router.websocket("/api/chat/exchange/{exchange_id}/stream")
+async def chat_exchange_stream(
+    exchange_id: DecodedDatabaseIdField,
+    websocket: WebSocket,
+) -> None:
+    app = get_app()
+    user = _get_websocket_user(websocket, app)
+    if user is None:
+        await websocket.close(code=status.WS_1008_POLICY_VIOLATION)
+        return
+
+    execution_manager = app[DataAnalysisChatExecutionManager]
+    trans = WorkRequestContext(app=app, user=user)
+    try:
+        await execution_manager.handle_exchange_stream(exchange_id, websocket, trans)
+    except ObjectNotFound:
+        await websocket.close(code=status.WS_1008_POLICY_VIOLATION)
